src/main/java/org/kiwi/crypto/api/CoinMarketCapRepository.java
package org.kiwi.crypto.api;

import com.google.inject.Inject;
import com.google.inject.name.Named;
import java.util.Collection;
import org.kiwi.crypto.currency.Currency;
import org.kiwi.crypto.currency.CurrencyMapper;
import org.kiwi.rest.RestClient;
import org.kiwi.rest.Unirest;

public class CoinMarketCapRepository implements CurrencyRepository {

    private final RestClient restClient;
    private final CurrencyMapper<String> currencyMapper;
    private final String endpoint;

    @Inject
    CoinMarketCapRepository(@Unirest RestClient restClient, @CoinMarketCap CurrencyMapper<String> currencyMapper,
            @Named(value = "crypto.coin.market.cap.endpoint") String endpoint) {
        this.restClient = restClient;
        this.currencyMapper = currencyMapper;
        this.endpoint = endpoint;
    }

    @Override
    public Collection<Currency> retrieveCurrencies() {
        try {
            String response = restClient.getGetResponseFrom(endpoint);
            return currencyMapper.map(response);
        } catch (Exception e) {
            throw new RuntimeException("Failed to load crypto currencies with " + this.getClass().getSimpleName(), e);
        }
    }
}
